ParquetStreamReader to cdm_reader_mapper.common.iterators
